인증 서버와 연동

+++ routes/v1/posts/controller.js
@@ -1,6 +1,6 @@
 const { Post } = require('../../../models');
 const { createResponse } = require('../../utils/response');
-const { POST_NOT_FOUND } = require('../../../errors');
+const { POST_NOT_FOUND, FORBIDDEN } = require('../../../errors');
 
 exports.getPosts = getPosts;
 exports.getPost = getPost;
@@ -21,7 +21,7 @@ async function getPost(req, res, next) {
   const { id } = req.params;
 
   try {
-    const doc = await Post.findById(id);
+    const doc = await Post.findById(id).populate({ path: 'writer', select: 'name -_id' });
     if (!doc) return next(POST_NOT_FOUND);
     res.json(createResponse(res, doc));
   } catch (e) {
@@ -31,7 +31,9 @@ async function getPost(req, res, next) {
 }
 
 async function createPost(req, res, next) {
-  const { body } = req;
+  const { body, user } = req;
+
+  body.writer = user.info;
 
   try {
     const doc = await Post.create(body);
@@ -42,11 +44,12 @@ async function createPost(req, res, next) {
 }
 
 async function updatePost(req, res, next) {
-  const { body: $set, params: { id } } = req;
+  const { body: $set, params: { id }, user } = req;
 
   try {
     const doc = await Post.findById(id);
     if (!doc) return next(POST_NOT_FOUND);
+    if (String(doc.writer) !== String(user.info)) return next(FORBIDDEN);
     await doc.updateOne({ $set });
     res.json(createResponse(res));
   } catch (e) {
@@ -55,11 +58,12 @@ async function updatePost(req, res, next) {
 }
 
 async function removePost(req, res, next) {
-  const { id } = req.params;
+  const { params: { id }, user } = req;
 
   try {
     const doc = await Post.findById(id);
     if (!doc) return next(POST_NOT_FOUND);
+    if (String(doc.writer) !== String(user.info)) return next(FORBIDDEN);
     await doc.deleteOne();
     res.json(createResponse(res));
   } catch (e) {
